Cleaned up release tests
