+++ internal/testutils/setup.go
@@ -1,0 +1,16 @@
+package testutils
+
+import (
+	"io"
+	"testing"
+
+	"github.com/ava-labs/avalanche-cli/pkg/ux"
+	"github.com/ava-labs/avalanchego/utils/logging"
+	"github.com/stretchr/testify/assert"
+)
+
+func SetupTest(t *testing.T) *assert.Assertions {
+	// use io.Discard to not print anything
+	ux.NewUserLog(logging.NoLog{}, io.Discard)
+	return assert.New(t)
+}
