+++ pkg/binutils/release_test.go
@@ -1,259 +1,37 @@
 package binutils
 
 import (
-	"archive/tar"
-	"archive/zip"
-	"bytes"
-	"compress/gzip"
-	"fmt"
-	"io"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	"github.com/ava-labs/avalanche-cli/internal/mocks"
+	"github.com/ava-labs/avalanche-cli/internal/testutils"
 	"github.com/ava-labs/avalanche-cli/pkg/application"
 	"github.com/ava-labs/avalanche-cli/pkg/config"
 	"github.com/ava-labs/avalanche-cli/pkg/prompts"
-	"github.com/ava-labs/avalanche-cli/pkg/ux"
 	"github.com/ava-labs/avalanchego/utils/logging"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 )
 
 const (
-	avalanchegoBin = "avalanchego"
-	pluginDirName  = "plugins"
-	evmBin         = "evm"
-	buildDirName   = "build"
-	subnetEVMBin   = "subnet-evm"
-	readme         = "README.md"
-	license        = "LICENSE"
-
 	version1 = "v1.17.1"
 	version2 = "v1.18.1"
+
+	avalanchegoBin = "avalanchego"
+	subnetEVMBin   = "subnet-evm"
 )
 
 var (
-	binary1   = []byte{0xde, 0xad, 0xbe, 0xef}
-	binary2   = []byte{0xfe, 0xed, 0xc0, 0xde}
-	evmBinary = []byte{0x00, 0xe1, 0x40, 0x00}
-
-	readmeContents  = []byte("README")
-	licenseContents = []byte("LICENSE")
+	binary1 = []byte{0xde, 0xad, 0xbe, 0xef}
+	binary2 = []byte{0xfe, 0xed, 0xc0, 0xde}
 )
-
-func setupTest(t *testing.T) *assert.Assertions {
-	// use io.Discard to not print anything
-	ux.NewUserLog(logging.NoLog{}, io.Discard)
-	return assert.New(t)
-}
-
-func verifyAvagoTarContents(assert *assert.Assertions, tarBytes []byte, version string) {
-	topDir := avalanchegoBinPrefix + version
-	bin := filepath.Join(topDir, avalanchegoBin)
-	plugins := filepath.Join(topDir, pluginDirName)
-	evm := filepath.Join(plugins, evmBin)
-
-	// topDirExists := false
-	binExists := false
-	pluginsExists := false
-	evmExists := false
-
-	file := bytes.NewReader(tarBytes)
-	gzRead, err := gzip.NewReader(file)
-	assert.NoError(err)
-	tarReader := tar.NewReader(gzRead)
-	assert.NoError(err)
-	for {
-		file, err := tarReader.Next()
-		if err == io.EOF {
-			break
-		}
-		assert.NoError(err)
-		fmt.Println("Archive contains:", file.Name)
-		switch file.Name {
-		case topDir:
-			// topDirExists = true
-			continue
-		case bin:
-			binExists = true
-		case plugins:
-			pluginsExists = true
-		case evm:
-			evmExists = true
-		default:
-			assert.FailNow("Tar has extra files")
-		}
-	}
-	// assert.True(topDirExists)
-	assert.True(binExists)
-	assert.True(pluginsExists)
-	assert.True(evmExists)
-}
-
-func verifySubnetEVMTarContents(assert *assert.Assertions, tarBytes []byte) {
-	binExists := false
-	readmeExists := false
-	licenseExists := false
-
-	file := bytes.NewReader(tarBytes)
-	gzRead, err := gzip.NewReader(file)
-	assert.NoError(err)
-	tarReader := tar.NewReader(gzRead)
-	assert.NoError(err)
-	for {
-		file, err := tarReader.Next()
-		if err == io.EOF {
-			break
-		}
-		assert.NoError(err)
-		fmt.Println("Archive contains:", file.Name)
-		switch file.Name {
-		case subnetEVMBin:
-			binExists = true
-		case readme:
-			readmeExists = true
-		case license:
-			licenseExists = true
-		default:
-			assert.FailNow("Tar has extra files: " + file.Name)
-		}
-	}
-	assert.True(binExists)
-	assert.True(readmeExists)
-	assert.True(licenseExists)
-}
-
-func verifyAvagoZipContents(assert *assert.Assertions, zipFile string) {
-	topDir := buildDirName
-	bin := filepath.Join(topDir, avalanchegoBin)
-	plugins := filepath.Join(topDir, pluginDirName)
-	evm := filepath.Join(plugins, evmBin)
-
-	topDirExists := false
-	binExists := false
-	pluginsExists := false
-	evmExists := false
-
-	reader, err := zip.OpenReader(zipFile)
-	assert.NoError(err)
-	defer reader.Close()
-	for _, file := range reader.File {
-		fmt.Println("Archive contains:", file.Name)
-		// Zip directories end in "/" which is annoying for string matching
-		switch strings.TrimSuffix(file.Name, "/") {
-		case topDir:
-			topDirExists = true
-		case bin:
-			binExists = true
-		case plugins:
-			pluginsExists = true
-		case evm:
-			evmExists = true
-		default:
-			assert.FailNow("Zip has extra files: " + file.Name)
-		}
-	}
-	assert.True(topDirExists)
-	assert.True(binExists)
-	assert.True(pluginsExists)
-	assert.True(evmExists)
-}
-
-func createDummyAvagoZip(assert *assert.Assertions, binary []byte) []byte {
-	sourceDir, err := os.MkdirTemp(os.TempDir(), "binutils-source")
-	assert.NoError(err)
-	defer os.RemoveAll(sourceDir)
-
-	topDir := filepath.Join(sourceDir, buildDirName)
-	err = os.Mkdir(topDir, 0o700)
-	assert.NoError(err)
-
-	binPath := filepath.Join(topDir, avalanchegoBin)
-	err = os.WriteFile(binPath, binary, 0o600)
-	assert.NoError(err)
-
-	pluginDir := filepath.Join(topDir, pluginDirName)
-	err = os.Mkdir(pluginDir, 0o700)
-	assert.NoError(err)
-
-	evmBinPath := filepath.Join(pluginDir, evmBin)
-	err = os.WriteFile(evmBinPath, evmBinary, 0o600)
-	assert.NoError(err)
-
-	// Put into zip
-	zipFile := "/tmp/avago.zip"
-	createZip(assert, topDir, zipFile)
-
-	verifyAvagoZipContents(assert, zipFile)
-
-	zipBytes, err := os.ReadFile(zipFile)
-	assert.NoError(err)
-	return zipBytes
-}
-
-func createDummyAvagoTar(assert *assert.Assertions, binary []byte, version string) []byte {
-	sourceDir, err := os.MkdirTemp(os.TempDir(), "binutils-source")
-	assert.NoError(err)
-	defer os.RemoveAll(sourceDir)
-
-	topDir := filepath.Join(sourceDir, avalanchegoBinPrefix+version)
-	err = os.Mkdir(topDir, 0o700)
-	assert.NoError(err)
-
-	binPath := filepath.Join(topDir, avalanchegoBin)
-	err = os.WriteFile(binPath, binary, 0o600)
-	assert.NoError(err)
-
-	pluginDir := filepath.Join(topDir, pluginDirName)
-	err = os.Mkdir(pluginDir, 0o700)
-	assert.NoError(err)
-
-	evmBinPath := filepath.Join(pluginDir, evmBin)
-	err = os.WriteFile(evmBinPath, evmBinary, 0o600)
-	assert.NoError(err)
-
-	// Put into tar
-	tarFile := "/tmp/avago.tar.gz"
-	createTarGz(assert, topDir, tarFile, true)
-	tarBytes, err := os.ReadFile(tarFile)
-	assert.NoError(err)
-	verifyAvagoTarContents(assert, tarBytes, version)
-	return tarBytes
-}
-
-func createDummySubnetEVMTar(assert *assert.Assertions, binary []byte) []byte {
-	sourceDir, err := os.MkdirTemp(os.TempDir(), "binutils-source")
-	assert.NoError(err)
-	defer os.RemoveAll(sourceDir)
-
-	binPath := filepath.Join(sourceDir, subnetEVMBin)
-	err = os.WriteFile(binPath, binary, 0o600)
-	assert.NoError(err)
-
-	readmePath := filepath.Join(sourceDir, readme)
-	err = os.WriteFile(readmePath, readmeContents, 0o600)
-	assert.NoError(err)
-
-	licensePath := filepath.Join(sourceDir, license)
-	err = os.WriteFile(licensePath, licenseContents, 0o600)
-	assert.NoError(err)
-
-	// Put into tar
-	tarFile := "/tmp/avago.tar.gz"
-	createTarGz(assert, sourceDir, tarFile, false)
-	tarBytes, err := os.ReadFile(tarFile)
-	assert.NoError(err)
-	verifySubnetEVMTarContents(assert, tarBytes)
-	return tarBytes
-}
 
 func setupInstallDir(assert *assert.Assertions) *application.Avalanche {
 	rootDir, err := os.MkdirTemp(os.TempDir(), "binutils-tests")
 	assert.NoError(err)
-	// defer os.RemoveAll(rootDir)
+	defer os.RemoveAll(rootDir)
 
 	app := application.New()
 	app.Setup(rootDir, logging.NoLog{}, &config.Config{}, prompts.NewPrompter())
@@ -261,9 +39,9 @@ func setupInstallDir(assert *assert.Assertions) *application.Avalanche {
 }
 
 func Test_installAvalancheGoWithVersion_Zip(t *testing.T) {
-	assert := setupTest(t)
+	assert := testutils.SetupTest(t)
 
-	zipBytes := createDummyAvagoZip(assert, binary1)
+	zipBytes := testutils.CreateDummyAvagoZip(assert, binary1)
 	app := setupInstallDir(assert)
 
 	mockInstaller := &mocks.Installer{}
@@ -286,9 +64,9 @@ func Test_installAvalancheGoWithVersion_Zip(t *testing.T) {
 }
 
 func Test_installAvalancheGoWithVersion_Tar(t *testing.T) {
-	assert := setupTest(t)
+	assert := testutils.SetupTest(t)
 
-	tarBytes := createDummyAvagoTar(assert, binary1, version1)
+	tarBytes := testutils.CreateDummyAvagoTar(assert, binary1, version1)
 
 	app := setupInstallDir(assert)
 
@@ -312,10 +90,10 @@ func Test_installAvalancheGoWithVersion_Tar(t *testing.T) {
 }
 
 func Test_installAvalancheGoWithVersion_MultipleCoinstalls(t *testing.T) {
-	assert := setupTest(t)
+	assert := testutils.SetupTest(t)
 
-	zipBytes1 := createDummyAvagoZip(assert, binary1)
-	zipBytes2 := createDummyAvagoZip(assert, binary2)
+	zipBytes1 := testutils.CreateDummyAvagoZip(assert, binary1)
+	zipBytes2 := testutils.CreateDummyAvagoZip(assert, binary2)
 	app := setupInstallDir(assert)
 
 	mockInstaller := &mocks.Installer{}
@@ -353,9 +131,9 @@ func Test_installAvalancheGoWithVersion_MultipleCoinstalls(t *testing.T) {
 }
 
 func Test_installSubnetEVMWithVersion(t *testing.T) {
-	assert := setupTest(t)
+	assert := testutils.SetupTest(t)
 
-	tarBytes := createDummySubnetEVMTar(assert, binary1)
+	tarBytes := testutils.CreateDummySubnetEVMTar(assert, binary1)
 	app := setupInstallDir(assert)
 
 	mockInstaller := &mocks.Installer{}
@@ -380,10 +158,10 @@ func Test_installSubnetEVMWithVersion(t *testing.T) {
 }
 
 func Test_installSubnetEVMWithVersion_MultipleCoinstalls(t *testing.T) {
-	assert := setupTest(t)
+	assert := testutils.SetupTest(t)
 
-	tarBytes1 := createDummySubnetEVMTar(assert, binary1)
-	tarBytes2 := createDummySubnetEVMTar(assert, binary2)
+	tarBytes1 := testutils.CreateDummySubnetEVMTar(assert, binary1)
+	tarBytes2 := testutils.CreateDummySubnetEVMTar(assert, binary2)
 	app := setupInstallDir(assert)
 
 	mockInstaller := &mocks.Installer{}
